Save before experimenting with traj

+++ drive_functions.h
@@ -9,7 +9,7 @@
 
   extern ODriveArduino odrive_array[];
   extern HardwareSerial *odrive_serial_array[];
-  extern double zero_offset_array[];
+  extern float zero_offset_array[];
   extern const float horizontal_offset_angle;
   extern const int gear_ratio;
